requirements.txt
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg[binary]>=3.2,<4.0
pydantic>=2.4,<3.0
pydantic-settings>=2.4,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
